First commit.
Adding two targets file for supporting the equiv of 'make dist'.
One is to extend Microsoft.Common.targets, and add a 'GetDistFiles'
target to all the msbuild project files in the solution.

The other, extensions the .sln file, and adds a 'Dist' target,
which collects the list of files from the projects by invoking
'GetDistFiles' on them. It copies all these into $(DistDir).

This is the first attempt, and it requires ToolsVersion=4.0.